Stream tweets

+++ src/db/entities.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import Column, Integer, String, BigInteger
 from sqlalchemy.ext.declarative import declarative_base
 from dotenv import load_dotenv
 
@@ -12,12 +12,14 @@ Base = declarative_base()
 class TwitterStreams(Base):
     __tablename__ = 'twitter_streams'
     domain_id = Column(Integer)
-    entity_id = Column(Integer)
+    entity_id = Column(BigInteger)
     entity_name = Column(String)
     language = Column(String)
     text = Column(String)
-    tweet_id = Column(Integer, primary_key=True, autoincrement=False)
+    tweet_id = Column(BigInteger, primary_key=True, autoincrement=False)
     # TODO: Add datetime column
+    # TODO: Add serial id
+
 
 if __name__ == "__main__":
